Change widget width and fix bug
