+++ client/src/components/RatingsNReviews/RatingsNReviews.jsx
@@ -3,9 +3,7 @@ import ReviewList from './ReviewList.jsx';
 import NewReview from './NewReview.jsx';
 const RatingsNReviews = ( {handleAverageRate, handleReviews, productId, currentProduct}) =>{
   return (
-    <div className= "RatingsNReviewsSection">
-      <ReviewList handleAverageRate={handleAverageRate} handleReviews={handleReviews} currentProduct={currentProduct} productId= {productId}/>
-    </div>
+    <ReviewList handleAverageRate={handleAverageRate} handleReviews={handleReviews} currentProduct={currentProduct} productId= {productId}/>
   );
 };
